requirements.txt
colorlog==6.7.0
telethon==1.39.0
cryptg==0.5.0.post0
flask==3.0.2
aiohttp==3.9.3
python-dotenv==1.0.1
TgCrypto==1.2.5